fix(web): extract preview panel hook to clear giant-component warning
Keep ArtifactAttachmentPreview thin; panel orchestration lives in a
hook so react-doctor/no-giant-component stays clean.
